third_party/WebKit/LayoutTests/media/controls-volume-slider-expected.txt

EXPECTED (audio.volume == '1') OK
EVENT(loadedmetadata)
EXPECTED (audio.volume > '0.4') OK
EXPECTED (audio.volume < '0.6') OK
EVENT(volumechange)
END OF TEST
